+++ client/src/core/plugins/modals/plugin.ts
@@ -1,10 +1,9 @@
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { createApp, inject, nextTick } from "vue";
 import type { InjectionKey, Plugin } from "vue";
 
 import PluginContainer from "../../components/PluginContainer.vue";
 
-import { useAssetPicker } from "./assetPicker";
-import type { AssetPickerFunction } from "./assetPicker";
 import { useConfirm } from "./confirm";
 import type { ConfirmFunction } from "./confirm";
 import { usePrompt } from "./prompt";
@@ -15,20 +14,21 @@ import type { SelectionBoxFunction } from "./selectionBox";
 const modalSymbol: InjectionKey<Modals> = Symbol("PlanarAllyModals");
 
 interface Modals {
-    assetPicker: AssetPickerFunction;
     confirm: ConfirmFunction;
     prompt: PromptFunction;
     selectionBox: SelectionBoxFunction;
 }
 
 async function createModals(): Promise<Modals> {
-    const assetPicker = useAssetPicker();
     const confirm = useConfirm();
     const prompt = usePrompt();
     const selectionBox = useSelectionBox();
 
     await nextTick(() => {
-        const app = createApp(PluginContainer, { assetPicker, confirm, prompt, selectionBox });
+        const app = createApp(PluginContainer, { confirm, prompt, selectionBox }).component(
+            "font-awesome-icon",
+            FontAwesomeIcon,
+        );
         const d = document.createElement("div");
         document.body.appendChild(d);
         app.mount(d);
@@ -36,7 +36,6 @@ async function createModals(): Promise<Modals> {
     });
 
     return {
-        assetPicker: assetPicker.show,
         confirm: confirm.show,
         prompt: prompt.ask,
         selectionBox: selectionBox.open,